Add person location - Add LocationController - Add button on person to show the map of where the person is and where it has been
